% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_get.R
\name{adobe_get}
\alias{adobe_get}
\title{adobe_get}
\usage{
adobe_get(datacenter, endpoint, auth)
}
\arguments{
\item{datacenter}{datacenter}

\item{endpoint}{endpoint}

\item{auth}{auth}
}
\value{
character
}
\description{
adobe_get
}
\examples{

}
