% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobe_oauth.R
\name{adobe_oauth}
\alias{adobe_oauth}
\title{adobe_get}
\usage{
adobe_oauth(key, secret,
  scope = "openid,AdobeID,read_organizations,additional_info.job_function,additional_info.projectedProductContext",
  outfile = ".httr-oauth")
}
\arguments{
\item{key}{key}

\item{secret}{secret}

\item{scope}{scope}

\item{outfile}{outfile}
}
\value{
Environment
}
\description{
adobe_get
}
\examples{

}
