% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdobeOAuth.R
\name{AdobeOAuth}
\alias{AdobeOAuth}
\title{OAuth2 authentication for Adobe Analytics API}
\usage{
AdobeOAuth(key, secret, globalCompanyId = NULL, scope = NULL,
  authfile = ".httr-oauth")
}
\arguments{
\item{key}{key}

\item{secret}{secret}

\item{globalCompanyId}{globalCompanyId}

\item{scope}{scope}

\item{authfile}{authfile}
}
\value{
Token 2.0 R6 Class (httr)
}
\description{
Description
}
\details{
Details
}
\examples{

}
