% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_AnalyticsDimension.R
\docType{data}
\name{AnalyticsDimension}
\alias{AnalyticsDimension}
\title{AnalyticsDimension}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalyticsDimension
}
\value{
AnalyticsDimension
}
\description{
AnalyticsDimension
}
\examples{

}
\keyword{datasets}
