% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticsMetric.R
\docType{data}
\name{AnalyticsMetric}
\alias{AnalyticsMetric}
\title{AnalyticsMetric}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalyticsMetric
}
\value{
AnalyticsMetric
}
\description{
AnalyticsMetric
}
\examples{

}
\keyword{datasets}
