% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticsSegmentResponseItem.R
\docType{data}
\name{AnalyticsSegmentResponseItem}
\alias{AnalyticsSegmentResponseItem}
\title{AnalyticsSegmentResponseItem}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalyticsSegmentResponseItem
}
\value{
AnalyticsSegmentResponseItem
}
\description{
AnalyticsSegmentResponseItem
}
\examples{

}
\keyword{datasets}
