% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_AnalyticsUser.R
\docType{data}
\name{AnalyticsUser}
\alias{AnalyticsUser}
\title{AnalyticsUser}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalyticsUser
}
\value{
AnalyticsUser
}
\description{
AnalyticsUser
}
\examples{

}
\keyword{datasets}
