% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_CalendarType.R
\docType{data}
\name{CalendarType}
\alias{CalendarType}
\title{CalendarType}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CalendarType
}
\value{
CalendarType
}
\description{
CalendarType
}
\examples{

}
\keyword{datasets}
