% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Favorite.R
\docType{data}
\name{Favorite}
\alias{Favorite}
\title{Favorite}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Favorite
}
\value{
Favorite
}
\description{
Favorite
}
\examples{

}
\keyword{datasets}
