% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimensions.R
\name{GetDimensions}
\alias{GetDimensions}
\title{GetDimensions}
\usage{
GetDimensions(rsid, dimension = NULL, as.data.frame = TRUE)
}
\arguments{
\item{rsid}{(character) The report suite ID}

\item{dimension}{(character) The dimension ID}

\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 (Dimensions | Dimension)
}
\description{
GetDimensions
}
\examples{
\dontrun{
dims <- GetDimensions("zwitchdev")
dims.nodf <- GetDimensions("zwitchdev", as.data.frame = FALSE)
dimsevar1 <- GetDimensions("zwitchdev", "evar1")
dimsevar1.nodf <- GetDimensions("zwitchdev", "evar1", as.data.frame = FALSE)

}
}
