% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetrics.R
\name{GetMetrics}
\alias{GetMetrics}
\title{Get metrics (custom events and pre-defined) defined in a given report suite(s)}
\usage{
GetMetrics(rsid, as.data.frame = TRUE, locale = NULL,
  segmentable = NULL, expansion = NULL)
}
\arguments{
\item{rsid}{(character) ID of desired report suite}

\item{as.data.frame}{(logical) Return result as data.frame}

\item{locale}{(character) Locale for encoding/localized spelling}

\item{segmentable}{(logical) Filter the metrics by if they are valid in a segment}

\item{expansion}{(character) Add extra metadata to items (comma-delimited list)}
}
\value{
data.frame or S3 'Metrics'
}
\description{
Get metrics (custom events and pre-defined) defined in a given report suite(s)
}
\examples{
\dontrun{

metrics <- GetMetrics("zwitchdev")
metrics.nodf <- GetMetrics("zwitchdev", as.data.frame = FALSE)

}
}
