% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportSuites.R
\name{GetReportSuites}
\alias{GetReportSuites}
\title{GetReportSuites}
\usage{
GetReportSuites(rsid = NULL, as.data.frame = TRUE, rsids = NULL,
  rsidContains = NULL, limit = 100, page = 0, expansion = NULL)
}
\arguments{
\item{rsid}{(character) ID of desired report suite}

\item{as.data.frame}{(logical) Return result as data.frame}

\item{rsids}{(character) Only include suites in this RSID list
(comma-delimited)}

\item{rsidContains}{(character) Only include suites whose rsid contains
rsidContains}

\item{limit}{(integer) Number of results per page}

\item{page}{(integer) Page number (base 0 - first page is "0")}

\item{expansion}{(character) Comma-delimited list of additional metadata
fields to include on response}
}
\value{
data.frame or S3 (ReportSuites | ReportSuite)
}
\description{
GetReportSuites
}
\examples{
\dontrun{
grs <- GetReportSuites()
grs.nodf <- GetReportSuites(as.data.frame = FALSE)
grsrsid <- GetReportSuites("zwitchdev")
grsrsid.nodf <- GetReportSuites("zwitchdev", as.data.frame = FALSE)

}
}
