% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSegments.R
\name{GetSegments}
\alias{GetSegments}
\title{GetSegments}
\usage{
GetSegments(id = NULL, as.data.frame = TRUE)
}
\arguments{
\item{id}{(character) Segment ID to retrieve}

\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 (Segments | Segment)
}
\description{
GetSegments
}
\examples{
\dontrun{
seg <- GetSegments()
seg.nodf <- GetSegments(as.data.frame = FALSE)
segid <- GetSegments("5433e4e6e4b02df70be4ac63")
segid.nodf <- GetSegments("5433e4e6e4b02df70be4ac63", FALSE)

}
}
