% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTags.R
\name{GetTags}
\alias{GetTags}
\title{GetTags}
\usage{
GetTags(id = NULL, as.data.frame = TRUE)
}
\arguments{
\item{id}{(character) ID for tag to retrieve}

\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 (Tags | Tag)
}
\description{
GetTags
}
\examples{
\dontrun{

gtag <- GetTags()
gtag.nodf <- GetTags(as.data.frame = FALSE) #returns S3 Tags
gtagid <- GetTags("28638")
gtagid.nodf <- GetTags("28638", as.data.frame = FALSE)
}
}
