% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUserById.R
\name{GetUserById}
\alias{GetUserById}
\title{GetUserById}
\usage{
GetUserById(id = "me")
}
\arguments{
\item{id}{User id to retrieve}
}
\value{
AnalyticsUser
}
\description{
GetUserById
}
\examples{
\dontrun{
gubi <- GetUserById()
}
}
