% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUserCompanyAccess.R
\name{GetUserCompanyAccess}
\alias{GetUserCompanyAccess}
\title{Companies user id can access}
\usage{
GetUserCompanyAccess(as.data.frame = TRUE)
}
\arguments{
\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 object of class "CompanyAccess"
}
\description{
Returns companies user can access, based on Token 2.0 credentials
}
\examples{
\dontrun{

df <- GetUserCompanyAccess()
discovery <- GetUserCompanyAccess(as.data.frame=FALSE)
}
}
\seealso{
\code{\link{AdobeOAuth}}
}
