% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuiteCollectionItem.R
\docType{data}
\name{SuiteCollectionItem}
\alias{SuiteCollectionItem}
\title{SuiteCollectionItem}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SuiteCollectionItem
}
\value{
SuiteCollectionItem
}
\description{
SuiteCollectionItem
}
\examples{

}
\keyword{datasets}
