% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedUsageItem.R
\docType{data}
\name{SummarizedUsageItem}
\alias{SummarizedUsageItem}
\title{SummarizedUsageItem}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummarizedUsageItem
}
\value{
SummarizedUsageItem
}
\description{
SummarizedUsageItem
}
\examples{

}
\keyword{datasets}
