% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_TaggedComponent.R
\docType{data}
\name{TaggedComponent}
\alias{TaggedComponent}
\title{TaggedComponent}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaggedComponent
}
\value{
TaggedComponent
}
\description{
TaggedComponent
}
\examples{

}
\keyword{datasets}
