% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdobeOAuth.R
\name{AdobeOAuth}
\alias{AdobeOAuth}
\title{OAuth2 authentication for Adobe Analytics API}
\usage{
AdobeOAuth(key, secret, globalCompanyId = NULL, scope = NULL,
  authfile = ".httr-oauth")
}
\arguments{
\item{key}{key}

\item{secret}{secret}

\item{globalCompanyId}{globalCompanyId}

\item{scope}{scope}

\item{authfile}{authfile}
}
\value{
Token 2.0 R6 Class (httr)
}
\description{
Description
}
\details{
Details
}
\examples{
\dontrun{
# get this from the adobe.io console
key <- "e6e798e57ba67330b42061e722c420cb"
secret <- "427bc850-714c-4bcg-857b-5b74e11206c6"

#this triggers the OAuth 3-legged authentication via browser
AdobeOAuth(key, secret)

}

}
