% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCalculatedMetrics.R
\name{GetCalculatedMetrics}
\alias{GetCalculatedMetrics}
\title{Get calculated metrics defined in a given report suite(s)}
\usage{
GetCalculatedMetrics(
  as.data.frame = TRUE,
  rsids = NULL,
  ownerId = NULL,
  calculatedMetricFilter = NULL,
  locale = NULL,
  name = NULL,
  tagNames = NULL,
  expansion = NULL
)
}
\arguments{
\item{as.data.frame}{(logical) Return result as data.frame}

\item{rsids}{(character) Only include calculated metrics tied to specified
RSID list (comma-delimited)}

\item{ownerId}{(integer) Only include calculated metrics owned by the
specified loginId}

\item{calculatedMetricFilter}{(character) Only include calculated metrics in
the specified list (comma-delimited list of IDs)}

\item{locale}{(character) Locale for encoding/localized spelling}

\item{name}{(character) Only include calculated metrics that contains the Name}

\item{tagNames}{(character) Only include calculated metrics that contains
one of the tags}

\item{expansion}{(character) Comma-delimited list of additional calculated
metric metadata fields to include on response}
}
\value{
data.frame or S3 'CalculatedMetrics'
}
\description{
Get calculated metrics defined in a given report suite(s)
}
\details{
The 'expansion' keyword allows for specifying a list of additional
  fields to return as part of the response. These fields should be specified
  as a character vector (string), separated with commas and no spaces between
  the fields.

  Because of the complex nature of having unknown combinations of extra
  fields returned on a function call, using the 'expansion' field with
  as.data.frame=TRUE can return undesirable results.
}
\examples{
\dontrun{
cm <- GetCalculatedMetrics()

cm.nodf <- GetCalculatedMetrics(as.data.frame=FALSE)

}
}
