% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCalculatedMetrics.R
\name{GetCalculatedMetrics}
\alias{GetCalculatedMetrics}
\title{GetCalculatedMetrics}
\usage{
GetCalculatedMetrics(globalCompanyId, id = NULL, as.data.frame = TRUE)
}
\arguments{
\item{as.data.frame}{}
}
\description{
GetCalculatedMetrics
}
