% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDateRanges.R
\name{GetDateRanges}
\alias{GetDateRanges}
\title{Get user-defined date ranges for a company account}
\usage{
GetDateRanges(as.data.frame = TRUE, locale = NULL,
  filterByIds = NULL, limit = 100, page = 0, expansion = NULL)
}
\arguments{
\item{as.data.frame}{(logical) Return result as data.frame}

\item{locale}{(character) Locale for encoding/localized spelling}

\item{filterByIds}{(character) Only include date ranges in the specified
list (comma-delimited list of IDs)}

\item{limit}{(integer) Number of results per page}

\item{page}{(integer) Page number (base 0 - first page is "0")}

\item{expansion}{(character) Comma-delimited list of additional date range
metadata fields to include on response}
}
\value{
data.frame or S3 'DateRanges'
}
\description{
Get user-defined date ranges for a company account
}
\details{
The 'expansion' keyword allows for specifying a list of additional
  fields to return as part of the response. These fields should be specified
  as a character vector (string), separated with commas and no spaces between
  the fields.

  Because of the complex nature of having unknown combinations of extra
  fields returned on a function call, using the 'expansion' field with
  as.data.frame=TRUE can return undesirable results.
}
\examples{
\dontrun{

gdr <- GetDateRanges()
gdr.nodf <- GetDateRanges(as.data.frame = FALSE)

}
}
