% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDateRanges.R
\name{GetDateRanges}
\alias{GetDateRanges}
\title{GetDateRanges}
\usage{
GetDateRanges(globalCompanyId, dateRangeId = NULL,
  as.data.frame = TRUE)
}
\arguments{
\item{as.data.frame}{}
}
\description{
GetDateRanges
}
