% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimensions.R
\name{GetDimensions}
\alias{GetDimensions}
\title{Get dimensions (props, eVars, etc.) defined in a given report suite(s)}
\usage{
GetDimensions(rsid, as.data.frame = TRUE, locale = NULL,
  segmentable = NULL, reportable = NULL, classifiable = NULL,
  expansion = NULL)
}
\arguments{
\item{rsid}{(character) The report suite ID}

\item{as.data.frame}{(logical) Return result as data.frame}

\item{locale}{(character) Locale for encoding/localized spelling}

\item{segmentable}{(logical) Only include dimensions that are valid within a
segment}

\item{reportable}{(logical) Only include dimensions that are valid within a
report}

\item{classifiable}{(logical) Only include classifiable dimensions}

\item{expansion}{(character) Add extra metadata to items (comma-delimited list)}
}
\value{
data.frame or S3 'Dimensions'
}
\description{
Get dimensions (props, eVars, etc.) defined in a given report suite(s)
}
\details{
The 'expansion' keyword allows for specifying a list of additional
  fields to return as part of the response. These fields should be specified
  as a character vector (string), separated with commas and no spaces between
  the fields.

  Because of the complex nature of having unknown combinations of extra
  fields returned on a function call, using the 'expansion' field with
  as.data.frame=TRUE can return undesirable results.
}
\examples{
\dontrun{
dims <- GetDimensions("zwitchdev")
dims.nodf <- GetDimensions("zwitchdev", as.data.frame = FALSE)

}
}
