% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetrics.R
\name{GetMetrics}
\alias{GetMetrics}
\title{GetMetrics}
\usage{
GetMetrics(rsid, metric = NULL, as.data.frame = TRUE, locale = NULL,
  segmentable = NULL, expansion = NULL)
}
\arguments{
\item{rsid}{(character) ID of desired report suite}

\item{metric}{(character) ID of the metric for which to retrieve info}

\item{as.data.frame}{(logical) Return result as data.frame}

\item{locale}{(character) Locale for encoding/localized spelling}

\item{segmentable}{(logical) Filter the metrics by if they are valid in a segment}

\item{expansion}{(character) Add extra metadata to items (comma-delimited list)}
}
\value{
data.frame or S3 (Metrics | Metric)
}
\description{
GetMetrics
}
\examples{
\dontrun{

metrics <- GetMetrics("zwitchdev")
metrics.nodf <- GetMetrics("zwitchdev", as.data.frame = FALSE)
metricspv <- GetMetrics("zwitchdev", "pageviews")
metricspv.nodf <- GetMetrics("zwitchdev", "pageviews", as.data.frame = FALSE)

}
}
