% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportSuites.R
\name{GetReportSuites}
\alias{GetReportSuites}
\title{GetReportSuites}
\usage{
GetReportSuites(rsid = NULL, as.data.frame = TRUE)
}
\arguments{
\item{rsid}{(character) ID of desired report suite}

\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 (ReportSuites | ReportSuite)
}
\description{
GetReportSuites
}
\examples{
\dontrun{
grs <- GetReportSuites()
grs.nodf <- GetReportSuites(as.data.frame = FALSE)
grsrsid <- GetReportSuites("zwitchdev")
grsrsid.nodf <- GetReportSuites("zwitchdev", as.data.frame = FALSE)

}
}
