% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSegments.R
\name{GetSegments}
\alias{GetSegments}
\title{Get user-defined segments for a company account}
\usage{
GetSegments(as.data.frame = TRUE, rsids = NULL, segmentFilter = NULL,
  locale = NULL, name = NULL, tagNames = NULL, limit = 100,
  page = 0, expansion = NULL)
}
\arguments{
\item{as.data.frame}{(logical) Return result as data.frame}

\item{rsids}{(character) Only include segments tied to specified RSID list
(comma-delimited)}

\item{segmentFilter}{(character) Only include segments in the specified list
(comma-delimited list of IDs)}

\item{locale}{(character) Locale for encoding/localized spelling}

\item{name}{(character) Only include segments that contains the name}

\item{tagNames}{(character) Only include segments that contains one of the tags}

\item{limit}{(integer) Number of results per page}

\item{page}{(integer) Page number (base 0 - first page is "0")}

\item{expansion}{(character) Comma-delimited list of additional metadata
fields to include on response}
}
\value{
data.frame or S3 'Segments'
}
\description{
Get user-defined segments for a company account
}
\examples{
\dontrun{
seg <- GetSegments()
seg.nodf <- GetSegments(as.data.frame = FALSE)

}
}
