% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTags.R
\name{GetTags}
\alias{GetTags}
\title{Get user-defined tags for a company account}
\usage{
GetTags(as.data.frame = TRUE)
}
\arguments{
\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 'Tags'
}
\description{
Get user-defined tags for a company account
}
\examples{
\dontrun{

gtag <- GetTags()
gtag.nodf <- GetTags(as.data.frame = FALSE) #returns S3 Tags
}
}
