% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUserCompanyAccess.R
\name{GetUserCompanyAccess}
\alias{GetUserCompanyAccess}
\title{Returns companies user can access}
\usage{
GetUserCompanyAccess(as.data.frame = TRUE)
}
\value{
data.frame or S3 object of class "CompanyAccess"
}
\description{
Returns companies user can access, based on the credentials returned from `AdobeOAuth()`
}
\examples{
\dontrun{

df <- GetUserCompanyAccess()
discovery <- GetUserCompanyAccess(as.data.frame=FALSE)
}
}
\seealso{
\code{\link{AdobeOAuth}}
}
