% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUsers.R
\name{GetUsers}
\alias{GetUsers}
\title{Get users for a company account}
\usage{
GetUsers(as.data.frame = TRUE, limit = 100, page = 0)
}
\arguments{
\item{as.data.frame}{(logical) Return result as data.frame}

\item{limit}{(integer) Number of results per page}

\item{page}{(integer) Page number (base 0 - first page is "0")}
}
\value{
data.frame or S3 'Users'
}
\description{
Get users for a company account
}
\examples{
\dontrun{

users <- GetUsers()
users.nodf <- GetUsers(as.data.frame = FALSE) #returns S3 Users

}
}
