% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUsers.R
\name{GetUsers}
\alias{GetUsers}
\title{GetUsers}
\usage{
GetUsers(id = NULL, as.data.frame = TRUE)
}
\arguments{
\item{id}{(character) ID of user you want to retrieve}

\item{as.data.frame}{(logical) Return result as data.frame}
}
\value{
data.frame or S3 (Users | User)
}
\description{
GetUsers
}
\examples{
\dontrun{

users <- GetUsers()
users.nodf <- GetUsers(as.data.frame = FALSE) #returns S3 Users
userme <- GetUsers("me")
userme.nodf <- GetUsers("me", as.data.frame = FALSE)

}
}
